# clients/packages/void-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Void API",
    "summary": "Void HTTP API",
    "version": "1.0.0",
    "description": "Void migration compatibility contract. Organization login, identity creation and navigation, and customer bindings are implemented in Polar. Other operations preserve the imported SDK API for later backend migration."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local Polar server with the Void API enabled"
    }
  ],
  "paths": {
...
    "/v1/void/events": {
      "get": {
        "tags": ["void", "private", "events"],
        "summary": "List Events",
        "description": "**Scopes**: `void:read` `void:write`",
        "operationId": "events:list",
        "security": [
          {
            "oat": ["void:read"]
          },
          {
            "oat": ["void:write"]
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only events with this name",
              "title": "Name"
            },
            "description": "Only events with this name"
          },
          {
            "name": "external_identity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this actor",
              "title": "External Identity Id"
            },
            "description": "Only this actor"
          },
          {
            "name": "external_root_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this tree's root",
              "title": "External Root Id"
            },
            "description": "Only this tree's root"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-polar-allowed-subjects": ["Organization"],
        "x-speakeasy-group": "events",
        "x-speakeasy-name-override": "list_events",
        "x-void-migration-status": "implemented"
      },
      "post": {
        "tags": ["void", "private", "events"],
        "summary": "Ingest",
        "description": "**Scopes**: `void:write`",
        "operationId": "events:ingest",
        "security": [
          {
            "oat": ["void:write"]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EventCreate"
                },
                "maxItems": 1000,
                "title": "Events"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsIngestResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAttribution"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservedEventName"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-polar-allowed-subjects": ["Organization"],
        "x-speakeasy-group": "events",
        "x-speakeasy-name-override": "ingest",
        "x-void-migration-status": "implemented"
      }
    },
    "/v1/void/reducers": {
      "get": {
        "tags": ["void", "private", "reducers"],
        "summary": "List Reducers",
        "description": "**Scopes**: `void:read` `void:write`",
        "operationId": "reducers:list",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Reducer"
                  },
                  "type": "array",
                  "title": "Response Reducers:List"
                }
              }
            }
          }
        },
        "security": [
          {
            "oat": ["void:read"]
          },
          {
            "oat": ["void:write"]
          }
        ],
        "x-polar-allowed-subjects": ["Organization"],
        "x-speakeasy-group": "reducers",
        "x-speakeasy-name-override": "list_reducers",
        "x-void-migration-status": "implemented"
      },
      "post": {
        "tags": ["void", "private", "reducers"],
        "summary": "Create Reducer",
        "description": "**Scopes**: `void:write`",
        "operationId": "reducers:create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReducerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reducer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidReducer"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlugTaken"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "oat": ["void:write"]
          }
        ],
        "x-polar-allowed-subjects": ["Organization"],
        "x-speakeasy-group": "reducers",
        "x-speakeasy-name-override": "create_reducer",
        "x-void-migration-status": "implemented"
      }
    },
    "/v1/void/reducers/{id}": {
      "get": {
        "tags": ["void", "private", "reducers"],
        "summary": "Get Reducer",
        "description": "**Scopes**: `void:read` `void:write`",
        "operationId": "reducers:get",
        "security": [
          {
            "oat": ["void:read"]
          },
          {
            "oat": ["void:write"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reducer"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFound"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-polar-allowed-subjects": ["Organization"],
        "x-speakeasy-group": "reducers",
        "x-speakeasy-name-override": "get_reducer",
        "x-void-migration-status": "implemented"
      }
    },
    "/v1/void/reducers/{id}/records": {
      "get": {
        "tags": ["void", "private", "reducers"],
        "summary": "Records",
        "description": "**Scopes**: `void:read` `void:write`",
        "operationId": "reducers:records",
        "security": [
          {
            "oat": ["void:read"]
          },
          {
            "oat": ["void:write"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "external_identity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "External Identity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReducerRecord"
                  },
                  "title": "Response Reducers:Records"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReducerNotDict"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFound"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-polar-allowed-subjects": ["Organization"],
        "x-speakeasy-group": "reducers",
        "x-speakeasy-name-override": "records",
        "x-void-migration-status": "implemented"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountAggregation": {
        "properties": {
          "func": {
            "type": "string",
            "const": "count",
            "title": "Func",
            "default": "count"
          }
        },
        "type": "object",
        "title": "CountAggregation"
      },
...
      "DerivedAggregation": {
        "properties": {
          "func": {
            "type": "string",
            "const": "derive",
            "title": "Func",
            "default": "derive"
          },
          "inputs": {
            "patternProperties": {
              "^[A-Za-z_][A-Za-z0-9_]*$": {
                "type": "string"
              }
            },
            "type": "object",
            "minProperties": 1,
            "title": "Inputs"
          },
          "expression": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Expression"
          }
        },
        "type": "object",
        "required": ["inputs", "expression"],
        "title": "DerivedAggregation"
      },
...
      "Event": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "examples": ["2026-01-01T00:00:00.000000Z"]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source": {
            "$ref": "#/components/schemas/EventSource"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "external_identity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Identity Id"
          },
          "external_root_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Root Id",
            "description": "The actor's root, stamped when the event was accepted."
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "timestamp",
          "name",
          "source",
          "external_id",
          "external_identity_id",
          "external_root_id",
          "metadata"
        ],
        "title": "Event"
      },
      "EventCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "external_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "External Id"
          },
          "external_identity_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "External Identity Id",
            "description": "An existing identity that performed the action. Usage rolls up to its root."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Event time with timezone, from 1970-01-01 inclusive to 2106-01-01 exclusive.",
            "examples": ["2026-01-01T00:00:00.000000Z"]
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": ["name", "external_id"],
        "title": "EventCreate"
      },
      "EventSource": {
        "type": "string",
        "enum": ["user", "system"],
        "title": "EventSource"
      },
      "EventsIngestResponse": {
        "properties": {
          "saved": {
            "type": "integer",
            "title": "Saved"
          },
          "ignored": {
            "type": "integer",
            "title": "Ignored"
          }
        },
        "type": "object",
        "required": ["saved", "ignored"],
        "title": "EventsIngestResponse"
      },
      "EventsList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": ["items", "pagination"],
        "title": "EventsList"
      },
      "Filter-Input": {
        "properties": {
          "conjunction": {
            "$ref": "#/components/schemas/FilterConjunction"
          },
          "clauses": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FilterClause"
                },
                {
                  "$ref": "#/components/schemas/Filter-Input"
                }
              ]
            },
            "type": "array",
            "title": "Clauses"
          }
        },
        "type": "object",
        "required": ["conjunction", "clauses"],
        "title": "Filter"
      },
      "Filter-Output": {
        "properties": {
          "conjunction": {
            "$ref": "#/components/schemas/FilterConjunction"
          },
          "clauses": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FilterClause"
                },
                {
                  "$ref": "#/components/schemas/Filter-Output"
                }
              ]
            },
            "type": "array",
            "title": "Clauses"
          }
        },
        "type": "object",
        "required": ["conjunction", "clauses"],
        "title": "Filter"
      },
      "FilterClause": {
        "properties": {
          "property": {
            "type": "string",
            "title": "Property"
          },
          "operator": {
            "$ref": "#/components/schemas/FilterOperator"
          },
          "value": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "integer",
                "maximum": 2147483647.0,
                "minimum": -2147483648.0
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": ["property", "operator", "value"],
        "title": "FilterClause"
      },
      "FilterConjunction": {
        "type": "string",
        "enum": ["and", "or"],
        "title": "FilterConjunction"
      },
      "FilterOperator": {
        "type": "string",
        "enum": ["eq", "ne", "gt", "gte", "lt", "lte", "like", "not_like"],
        "title": "FilterOperator"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "Pagination": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": ["total_count"],
        "title": "Pagination"
      },
...
      "PropertyAggregation": {
        "properties": {
          "func": {
            "type": "string",
            "enum": ["sum", "max", "min"],
            "title": "Func"
          },
          "property": {
            "type": "string",
            "title": "Property"
          }
        },
        "type": "object",
        "required": ["func", "property"],
        "title": "PropertyAggregation"
      },
      "RecordAggregation": {
        "properties": {
          "func": {
            "type": "string",
            "enum": ["first", "last"],
            "title": "Func"
          }
        },
        "type": "object",
        "required": ["func"],
        "title": "RecordAggregation"
      },
...
      "Reducer": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": ["2026-01-01T00:00:00.000000Z"]
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Filter-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "aggregation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CountAggregation"
              },
              {
                "$ref": "#/components/schemas/PropertyAggregation"
              },
              {
                "$ref": "#/components/schemas/RecordAggregation"
              },
              {
                "$ref": "#/components/schemas/DerivedAggregation"
              }
            ],
            "title": "Aggregation",
            "discriminator": {
              "propertyName": "func",
              "mapping": {
                "count": "#/components/schemas/CountAggregation",
                "derive": "#/components/schemas/DerivedAggregation",
                "first": "#/components/schemas/RecordAggregation",
                "last": "#/components/schemas/RecordAggregation",
                "max": "#/components/schemas/PropertyAggregation",
                "min": "#/components/schemas/PropertyAggregation",
                "sum": "#/components/schemas/PropertyAggregation"
              }
            }
          },
          "map": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map"
          },
          "type": {
            "type": "string",
            "enum": ["scalar", "dict"],
            "title": "Type",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "slug",
          "filter",
          "aggregation",
          "map",
          "type"
        ],
        "title": "Reducer"
      },
      "ReducerCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Slug"
          },
          "filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Filter-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "aggregation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CountAggregation"
              },
              {
                "$ref": "#/components/schemas/PropertyAggregation"
              },
              {
                "$ref": "#/components/schemas/RecordAggregation"
              },
              {
                "$ref": "#/components/schemas/DerivedAggregation"
              }
            ],
            "title": "Aggregation",
            "discriminator": {
              "propertyName": "func",
              "mapping": {
                "count": "#/components/schemas/CountAggregation",
                "derive": "#/components/schemas/DerivedAggregation",
                "first": "#/components/schemas/RecordAggregation",
                "last": "#/components/schemas/RecordAggregation",
                "max": "#/components/schemas/PropertyAggregation",
                "min": "#/components/schemas/PropertyAggregation",
                "sum": "#/components/schemas/PropertyAggregation"
              }
            }
          },
          "map": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map"
          }
        },
        "type": "object",
        "required": ["slug", "aggregation"],
        "title": "ReducerCreate"
      },
      "ReducerRecord": {
        "properties": {
          "external_identity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Identity Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "examples": ["2026-01-01T00:00:00.000000Z"]
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["external_identity_id", "timestamp", "data"],
        "title": "ReducerRecord"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
      "ResourceNotFound": {
        "properties": {
          "error": {
            "type": "string",
            "const": "ResourceNotFound",
            "title": "Error",
            "examples": ["ResourceNotFound"]
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["error", "detail"],
        "title": "ResourceNotFound"
      },
      "InvalidAttribution": {
        "properties": {
          "error": {
            "type": "string",
            "const": "InvalidAttribution",
            "title": "Error",
            "examples": ["InvalidAttribution"]
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["error", "detail"],
        "title": "InvalidAttribution"
      },
      "InvalidReducer": {
        "properties": {
          "error": {
            "type": "string",
            "const": "InvalidReducer",
            "title": "Error",
            "examples": ["InvalidReducer"]
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["error", "detail"],
        "title": "InvalidReducer"
      },
      "ReducerNotDict": {
        "properties": {
          "error": {
            "type": "string",
            "const": "ReducerNotDict",
            "title": "Error",
            "examples": ["ReducerNotDict"]
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["error", "detail"],
        "title": "ReducerNotDict"
      },
...
      "ReservedEventName": {
        "properties": {
          "error": {
            "type": "string",
            "const": "ReservedEventName",
            "title": "Error",
            "examples": ["ReservedEventName"]
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["error", "detail"],
        "title": "ReservedEventName"
      },
      "SlugTaken": {
        "properties": {
          "error": {
            "type": "string",
            "const": "SlugTaken",
            "title": "Error",
            "examples": ["SlugTaken"]
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["error", "detail"],
        "title": "SlugTaken"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "oat": {
        "type": "http",
        "description": "An organization access token with Void scopes. For local development, use scripts.generate_void_token for an allowlisted organization.",
        "scheme": "bearer"
      }
    }
  }
}